+++ .gitignore
@@ -5,5 +5,6 @@
 *.log
 
 /dev/*
+/cogs/palgame/*
 !gamedata/*.json
 /gamedata/kits.json
