+++ cogs/playerlist.py
@@ -1,10 +1,10 @@
 import nextcord
 from nextcord.ext import commands
 from nextcord.ui import View
-from util.rconutility import RconUtility
-import util.constants as constants
+from utils.rconutility import RconUtility
+import utils.constants as constants
 import datetime
-from util.database import get_server_details, server_autocomplete
+from utils.database import get_server_details, server_autocomplete
 
 class PlayerListView(View):
     def __init__(self, server, player_data):
